establish workflow examples
