+++ compute/slurm/podHandler.go
@@ -17,26 +17,16 @@ package slurm
 import (
 	"context"
 	"fmt"
-	"io/fs"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"text/template"
 
 	"github.com/carv-ics-forth/hpk/compute"
-	"github.com/carv-ics-forth/hpk/pkg/envvars"
-	"github.com/carv-ics-forth/hpk/pkg/fieldpath"
 	"github.com/go-logr/logr"
 	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
-	k8errors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/json"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -151,7 +141,7 @@ func DeletePod(podKey client.ObjectKey) error {
 
 	idType, podID := ParsePodID(pod)
 
-	if idType != IDTypeEmpty {
+	if idType != JobIDTypeEmpty {
 		out, err := CancelJob(podID)
 		if err != nil {
 			SystemError(err, "failed to cancel job '%s'. out: '%s'", podID, out)
@@ -237,7 +227,7 @@ func CreatePod(ctx context.Context, pod *corev1.Pod) error {
 	 *---------------------------------------------------*/
 	logger.Info(" * Creating Backend Volumes")
 
-	h.makeMounts(ctx)
+	h.prepareVolumes(ctx)
 
 	/*---------------------------------------------------
 	 * Set listeners for async changes on the VirtualEnvironment
@@ -258,7 +248,7 @@ func CreatePod(ctx context.Context, pod *corev1.Pod) error {
 	var initContainers []Process
 
 	for i, container := range pod.Spec.InitContainers {
-		job, err := h.makeApptainerCommands(&pod.Spec.InitContainers[i], true)
+		job, err := h.buildContainer(&pod.Spec.InitContainers[i], true)
 		if err != nil {
 			SystemError(err, "creation request failed for container '%s'", container.Name)
 		}
@@ -274,7 +264,7 @@ func CreatePod(ctx context.Context, pod *corev1.Pod) error {
 	var containers []Process
 
 	for i, container := range pod.Spec.Containers {
-		job, err := h.makeApptainerCommands(&pod.Spec.Containers[i], false)
+		job, err := h.buildContainer(&pod.Spec.Containers[i], false)
 		if err != nil {
 			return errors.Wrapf(err, "creation request failed for container '%s'", container.Name)
 		}
@@ -350,13 +340,13 @@ func CreatePod(ctx context.Context, pod *corev1.Pod) error {
 	 *---------------------------------------------------*/
 	logger.Info(" * Submit sbatch to Slurm", "scriptPath", job.VirtualEnv.ConstructorPath)
 
-	jobID, err := h.submitJob(sbatchScriptPath)
+	jobID, err := SubmitJob(sbatchScriptPath)
 	if err != nil {
 		SystemError(err, "failed to submit job")
 	}
 
 	/*-- Update Pod with the JobID --*/
-	SetPodID(h.Pod, IDTypeSlurm, jobID)
+	SetPodID(h.Pod, JobIDTypeSlurm, jobID)
 	if err != nil {
 		SystemError(err, "failed to submit job")
 	}
@@ -367,7 +357,7 @@ func CreatePod(ctx context.Context, pod *corev1.Pod) error {
 	return nil
 }
 
-func (h *podHandler) makeApptainerCommands(container *corev1.Container, isInitContainer bool) (Process, error) {
+func (h *podHandler) buildContainer(container *corev1.Container, isInitContainer bool) (Process, error) {
 	h.logger.Info(" == Apptainer ==", "container", container.Name)
 
 	containerPath := h.podDirectory.Container(container.Name)
@@ -454,443 +444,4 @@ func (h *podHandler) makeApptainerCommands(container *corev1.Container, isInitCo
 		StderrPath:   containerPath.StderrPath(),
 		ExitCodePath: containerPath.ExitCodePath(),
 	}, nil
-}
-
-func (h *podHandler) submitJob(scriptFilePath string) (string, error) {
-	out, err := SubmitJob(scriptFilePath)
-	if err != nil {
-		SystemError(err, "sbatch submission error. out : '%s'", out)
-	}
-
-	expectedOutput := regexp.MustCompile(`Submitted batch job (?P<jid>\d+)`)
-	jid := expectedOutput.FindStringSubmatch(out)
-
-	if _, err := strconv.Atoi(jid[1]); err != nil {
-		SystemError(err, "Cant convert jid as integer. Parsed irregular output!")
-	}
-	return jid[1], nil
-}
-
-func (h *podHandler) makeMounts(ctx context.Context) {
-	/*---------------------------------------------------
-	 * Copy volumes from local VirtualEnvironment to remote HPCBackend
-	 *---------------------------------------------------*/
-	for _, vol := range h.Pod.Spec.Volumes {
-		switch {
-		case vol.VolumeSource.ConfigMap != nil:
-			/*---------------------------------------------------
-			 * ConfigMap
-			 * Example:
-				volumes:
-				 - name: config-volume
-				   configMap:
-					 name: game-config
-			 *---------------------------------------------------*/
-			var configMap corev1.ConfigMap
-
-			source := vol.VolumeSource.ConfigMap
-
-			key := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: source.Name}
-
-			if err := compute.K8SClient.Get(ctx, key, &configMap); k8errors.IsNotFound(err) {
-				if source.Optional != nil && *source.Optional == false {
-					PodError(h.Pod, ReasonObjectNotFound, "configMap '%s'", key)
-				}
-
-				/*-- configMap is optional, and we can safely skip the step --*/
-				return
-			} else if err != nil {
-				SystemError(err, "error getting configMap '%s'", key)
-			}
-
-			// .hpk/namespace/podName/volName/*
-			podConfigMapDir, err := h.podDirectory.CreateSubDirectory(vol.Name)
-			if err != nil {
-				SystemError(err, "cannot create dir '%s' for configMap", podConfigMapDir)
-			}
-
-			for k, v := range configMap.Data {
-				// TODO: Ensure that these files are deleted in failure cases
-				fullPath := filepath.Join(podConfigMapDir, k)
-
-				if err := os.WriteFile(fullPath, []byte(v), fs.FileMode(*source.DefaultMode)); err != nil {
-					SystemError(err, "cannot write config map file '%s'", fullPath)
-				}
-			}
-
-		case vol.VolumeSource.Secret != nil:
-			/*---------------------------------------------------
-			 * Secret
-			 *---------------------------------------------------*/
-			var secret corev1.Secret
-
-			source := vol.VolumeSource.Secret
-
-			key := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: source.SecretName}
-
-			if err := compute.K8SClient.Get(ctx, key, &secret); k8errors.IsNotFound(err) {
-				if source.Optional != nil && *source.Optional == false {
-					PodError(h.Pod, ReasonObjectNotFound, "secret '%s'", key)
-				}
-
-				/*-- secret is optional, and we can safely skip the step --*/
-				return
-			} else if err != nil {
-				SystemError(err, "error getting secret '%s'", key)
-			}
-
-			// .hpk/namespace/podName/secretName/*
-			podSecretDir, err := h.podDirectory.CreateSubDirectory(vol.Name)
-			if err != nil {
-				SystemError(err, "cannot create dir '%s' for secrets", podSecretDir)
-			}
-
-			for k, v := range secret.Data {
-				fullPath := filepath.Join(podSecretDir, k)
-
-				if err := os.WriteFile(fullPath, v, fs.FileMode(*source.DefaultMode)); err != nil {
-					SystemError(err, "Could not write secret file %s", fullPath)
-				}
-			}
-
-		case vol.VolumeSource.EmptyDir != nil:
-			/*---------------------------------------------------
-			 * EmptyDir
-			 *---------------------------------------------------*/
-			emptyDir, err := h.podDirectory.CreateSubDirectory(vol.Name)
-			if err != nil {
-				SystemError(err, "cannot create dir '%s' for emptyDir", emptyDir)
-			}
-			// without size limit for now
-
-		case vol.VolumeSource.DownwardAPI != nil:
-			/*---------------------------------------------------
-			 * Downward API
-			 *---------------------------------------------------*/
-			downApiDir, err := h.podDirectory.CreateSubDirectory(vol.Name)
-			if err != nil {
-				SystemError(err, "cannot create dir '%s' for downwardApi", downApiDir)
-			}
-
-			for _, item := range vol.DownwardAPI.Items {
-				itemPath := filepath.Join(downApiDir, item.Path)
-				value, err := fieldpath.ExtractFieldPathAsString(h.Pod, item.FieldRef.FieldPath)
-				if err != nil {
-					PodError(h.Pod, ReasonSpecError, err.Error())
-
-					return
-				}
-
-				if err := os.WriteFile(itemPath, []byte(value), fs.FileMode(*vol.Projected.DefaultMode)); err != nil {
-					SystemError(err, "cannot write config map file '%s'", itemPath)
-				}
-			}
-
-		case vol.VolumeSource.HostPath != nil:
-			/*---------------------------------------------------
-			 * HostPath
-			 *---------------------------------------------------*/
-			switch *vol.VolumeSource.HostPath.Type {
-			case corev1.HostPathUnset:
-				// For backwards compatible, leave it empty if unset
-				if path, err := h.podDirectory.CreateSubDirectory(vol.Name); err != nil {
-					SystemError(err, "cannot create HostPathUnset at path '%s'", path)
-				}
-
-			case corev1.HostPathDirectoryOrCreate:
-				// If nothing exists at the given path, an empty directory will be created there
-				// as needed with file mode 0755, having the same group and ownership with Kubelet.
-				if path, err := h.podDirectory.CreateSubDirectory(vol.Name); err != nil {
-					SystemError(err, "cannot create HostPathDirectoryOrCreate at path '%s'", path)
-				}
-			case corev1.HostPathDirectory:
-				// A directory must exist at the given path
-				if path, err := h.podDirectory.CreateSubDirectory(vol.Name); err != nil {
-					SystemError(err, "cannot create HostPathDirectory at path '%s'", path)
-				}
-
-			case corev1.HostPathFileOrCreate:
-				// If nothing exists at the given path, an empty file will be created there
-				// as needed with file mode 0644, having the same group and ownership with Kubelet.
-				// .hpk/podName/volName/*
-				f, err := os.Create(string(h.podDirectory))
-				if err != nil {
-					SystemError(err, "cannot create  '%s'", string(h.podDirectory))
-				}
-
-				if err := f.Close(); err != nil {
-					SystemError(err, "cannot close file '%s'", string(h.podDirectory))
-				}
-			case corev1.HostPathFile:
-				// A file must exist at the given path
-				// .hpk/podName/volName/*
-				f, err := os.Create(string(h.podDirectory))
-				if err != nil {
-					SystemError(err, "cannot create '%s'", string(h.podDirectory))
-				}
-
-				if err := f.Close(); err != nil {
-					SystemError(err, "cannot close file '%s'", string(h.podDirectory))
-				}
-
-			case corev1.HostPathSocket, corev1.HostPathCharDev, corev1.HostPathBlockDev:
-				// A UNIX socket/char device/ block device must exist at the given path
-				continue
-			default:
-				panic("bug")
-			}
-
-		case vol.VolumeSource.Projected != nil:
-			/*---------------------------------------------------
-			 * Projected
-			 *---------------------------------------------------*/
-			projectedVolPath, err := h.podDirectory.CreateSubDirectory(vol.Name)
-			if err != nil {
-				SystemError(err, "cannot create dir '%s' for projected volume", projectedVolPath)
-			}
-
-			for _, projectedSrc := range vol.Projected.Sources {
-				switch {
-				case projectedSrc.DownwardAPI != nil:
-					/*---------------------------------------------------
-					 * Projected DownwardAPI
-					 *---------------------------------------------------*/
-					for _, item := range projectedSrc.DownwardAPI.Items {
-						itemPath := filepath.Join(projectedVolPath, item.Path)
-
-						value, err := fieldpath.ExtractFieldPathAsString(h.Pod, item.FieldRef.FieldPath)
-						if err != nil {
-							PodError(h.Pod, ReasonSpecError, err.Error())
-
-							return
-						}
-
-						if err := os.WriteFile(itemPath, []byte(value), fs.FileMode(*vol.Projected.DefaultMode)); err != nil {
-							SystemError(err, "cannot write config map file '%s'", itemPath)
-						}
-					}
-
-				case projectedSrc.ServiceAccountToken != nil:
-					/*---------------------------------------------------
-					 * Projected ServiceAccountToken
-					 *---------------------------------------------------*/
-					var serviceAccount corev1.ServiceAccount
-
-					key := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: h.Pod.Spec.ServiceAccountName}
-
-					if err := compute.K8SClient.Get(ctx, key, &serviceAccount); k8errors.IsNotFound(err) {
-						PodError(h.Pod, ReasonObjectNotFound, "serviceaccount '%s'", key)
-
-						return
-					} else if err != nil {
-						SystemError(err, "error getting serviceaccount '%s'", key)
-					}
-
-					/*
-						automount follows the instructions of:
-						https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/
-					*/
-
-					automount := true
-
-					switch {
-					case h.Pod.Spec.AutomountServiceAccountToken != nil:
-						automount = *h.Pod.Spec.AutomountServiceAccountToken
-					case serviceAccount.AutomountServiceAccountToken != nil:
-						automount = *serviceAccount.AutomountServiceAccountToken
-					}
-
-					if automount {
-						for _, secretRef := range serviceAccount.Secrets {
-							var secret corev1.Secret
-
-							secretKey := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: secretRef.Name}
-
-							if err := compute.K8SClient.Get(ctx, secretKey, &secret); k8errors.IsNotFound(err) {
-								PodError(h.Pod, ReasonObjectNotFound, "secret '%s'", secretKey)
-
-								return
-							} else if err != nil {
-								SystemError(err, "error getting secret '%s'", secretKey)
-							}
-
-							// TODO: Update upon exceeded expiration date
-							// TODO: Ensure that these files are deleted in failure cases
-							fullPath := filepath.Join(projectedVolPath, projectedSrc.ServiceAccountToken.Path)
-
-							if err := os.WriteFile(fullPath, secret.Data[projectedSrc.ServiceAccountToken.Path], fs.FileMode(0o766)); err != nil {
-								SystemError(err, "cannot write token '%s'", fullPath)
-							}
-						}
-					}
-
-					/*
-						source := projectedSrc.ServiceAccountToken
-
-						tokenRequest, err := compute.ClientSet.CoreV1().
-							ServiceAccounts(h.Pod.GetNamespace()).
-							CreateToken(ctx, h.Pod.Spec.ServiceAccountName, &authv1.TokenRequest{
-								Spec: authv1.TokenRequestSpec{
-									Audiences:         []string{source.Audience},
-									ExpirationSeconds: source.ExpirationSeconds,
-									BoundObjectRef: &authv1.BoundObjectReference{
-										Kind:       "Pod",
-										APIVersion: "v1",
-										Name:       h.Pod.GetName(),
-									},
-								},
-							}, metav1.CreateOptions{})
-
-						if err != nil {
-							return errors.Wrapf(err, "failed to create token for Pod '%s'", h.podKey)
-						}
-
-						// TODO: Update upon exceeded expiration date
-						// TODO: Ensure that these files are deleted in failure cases
-						fullPath := filepath.Join(projectedVolPath, projectedSrc.ServiceAccountToken.Path)
-
-						if err := os.WriteFile(fullPath, []byte(tokenRequest.Status.Token), fs.FileMode(0o766)); err != nil {
-							return errors.Wrapf(err, "cannot write token '%s'", fullPath)
-						}
-
-
-							var automount bool
-
-							switch {
-							case h.Pod.Spec.AutomountServiceAccountToken != nil:
-								automount = *h.Pod.Spec.AutomountServiceAccountToken
-							case serviceAccount.AutomountServiceAccountToken != nil:
-								automount = *serviceAccount.AutomountServiceAccountToken
-							}
-
-							if automount {
-								for _, secretRef := range serviceAccount.Secrets {
-									var secret corev1.Secret
-
-									secretKey := types.NamespacedName{Namespace: secretRef.Namespace, Name: secretRef.Name}
-
-									if err := compute.K8SClient.Get(ctx, secretKey, &secret); err != nil {
-										return errors.Wrapf(err, "failed to get secret '%s'", secretKey)
-									}
-
-									// TODO: Update upon exceeded expiration date
-									// TODO: Ensure that these files are deleted in failure cases
-									fullPath := filepath.Join(projectedVolPath, projectedSrc.ServiceAccountToken.Path)
-
-									if err := os.WriteFile(fullPath, secret.Data[projectedSrc.ServiceAccountToken.Path], fs.FileMode(0o766)); err != nil {
-										return errors.Wrapf(err, "cannot write config map file '%s'", fullPath)
-									}
-								}
-							}
-					*/
-
-				case projectedSrc.ConfigMap != nil:
-					/*---------------------------------------------------
-					 * Projected ConfigMap
-					 *---------------------------------------------------*/
-					var configmap corev1.ConfigMap
-
-					source := projectedSrc.ConfigMap
-
-					key := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: source.Name}
-
-					if err := compute.K8SClient.Get(ctx, key, &configmap); k8errors.IsNotFound(err) {
-						if source.Optional != nil && *source.Optional == false {
-							PodError(h.Pod, ReasonObjectNotFound, "configmap '%s'", key)
-						}
-
-						/*-- configmap is optional, and we can safely skip the step --*/
-						return
-					} else if err != nil {
-						SystemError(err, "error getting configmap '%s'", key)
-					}
-
-					for k, item := range configmap.Data {
-						// TODO: Ensure that these files are deleted in failure cases
-						itemPath := filepath.Join(projectedVolPath, k)
-						if err := os.WriteFile(itemPath, []byte(item), PodGlobalDirectoryPermissions); err != nil {
-							SystemError(err, "cannot write config map file '%s'", itemPath)
-						}
-					}
-
-				case projectedSrc.Secret != nil:
-					/*---------------------------------------------------
-					 * Projected Secret
-					 *---------------------------------------------------*/
-					var secret corev1.Secret
-
-					source := projectedSrc.Secret
-
-					key := types.NamespacedName{Namespace: h.Pod.GetNamespace(), Name: source.Name}
-
-					if err := compute.K8SClient.Get(ctx, key, &secret); k8errors.IsNotFound(err) {
-						if source.Optional != nil && *source.Optional == false {
-							PodError(h.Pod, ReasonObjectNotFound, "secret '%s'", key)
-						}
-
-						/*-- secret is optional, and we can safely skip the step --*/
-						return
-					} else if err != nil {
-						SystemError(err, "error getting secret '%s'", key)
-					}
-
-					for k, item := range secret.Data {
-						// TODO: Ensure that these files are deleted in failure cases
-						itemPath := filepath.Join(projectedVolPath, k)
-
-						if err := os.WriteFile(itemPath, item, PodGlobalDirectoryPermissions); err != nil {
-							SystemError(err, "cannot write config map file '%s'", itemPath)
-						}
-					}
-				default:
-					logrus.Warn(projectedSrc)
-
-					panic("Have I missed something ")
-				}
-			}
-		}
-	}
-}
-
-// getServiceEnvVarMap makes a map[string]string of env vars for services a  pod in namespace ns should see.
-// However, the kubelet implementation works on for Services with ClusterIPs.
-// In our case, the Services point directly to Pod IP's, and the implementation does not work.
-// The solution is to retrieve services from Endpoints.
-//
-// Original:
-// https://github.com/kubernetes/kubernetes/blob/1139bb177b2b35611c5ca16cc82f0e41a8bb107e/pkg/kubelet/kubelet_pods.go#L575
-func getServiceEnvVarMap(ctx context.Context, namespace string) ([]corev1.EnvVar, error) {
-	/*---------------------------------------------------
-	 * Get all service resources from master
-	 *---------------------------------------------------*/
-	var endpointsList corev1.EndpointsList
-
-	if err := compute.K8SClient.List(ctx, &endpointsList, &client.ListOptions{
-		LabelSelector: labels.Everything(),
-	}); err != nil {
-		return nil, errors.Wrap(err, "failed to list services when setting up env vars")
-	}
-
-	/*---------------------------------------------------
-	 * Populate services into service environment variables.
-	 *---------------------------------------------------*/
-	var mappedEndpoints []*corev1.Endpoints
-
-	for i, endpoint := range endpointsList.Items {
-		// ignore endpoints without IPs
-		if len(endpoint.Subsets) == 0 {
-			continue
-		}
-
-		// We always want to add environment variabled for master services
-		// from the master service namespace, even if enableServiceLinks is false.
-		// We also add environment variables for other services in the same
-		// namespace, if enableServiceLinks is true.
-		if endpoint.GetNamespace() == namespace ||
-			endpoint.GetNamespace() == metav1.NamespaceDefault {
-			mappedEndpoints = append(mappedEndpoints, &endpointsList.Items[i])
-		}
-	}
-
-	return envvars.FromEndpoints(mappedEndpoints), nil
 }
