+++ examples/services/nginx.yaml
