+++ examples/workflows/argo/independent-tasks.yaml
@@ -8,7 +8,7 @@
 apiVersion: argoproj.io/v1alpha1
 kind: Workflow
 metadata:
-  generateName: dag-diamond-
+  name: independent-tasks
   namespace: default
 spec:
   entrypoint: diamond
@@ -43,5 +43,3 @@ spec:
       container:
         image: alpine:3.7
         command: [ echo, "{{inputs.parameters.message}}" ]
-
-  dnsPolicy: ClusterFirst
